Move getGraphQLErrorsFromResult to internal
