+++ src/core/QueryManager.ts
@@ -42,13 +42,13 @@ import { print } from "@apollo/client/utilities";
 import { AutoCleanedWeakCache, cacheSizes } from "@apollo/client/utilities";
 import {
   DocumentTransform,
-  getGraphQLErrorsFromResult,
   graphQLResultHasError,
 } from "@apollo/client/utilities";
 import { __DEV__ } from "@apollo/client/utilities/environment";
 import {
   checkDocument,
   getDefaultValues,
+  getGraphQLErrorsFromResult,
   getOperationDefinition,
   getOperationName,
   hasClientExports,
