+++ src/utilities/common/errorHandling.ts
@@ -1,27 +1,10 @@
 import type { FetchResult } from "@apollo/client/link";
 import {
-  isExecutionPatchIncrementalResult,
+  getGraphQLErrorsFromResult,
   isNonEmptyArray,
 } from "@apollo/client/utilities/internal";
 
 export function graphQLResultHasError<T>(result: FetchResult<T>): boolean {
   const errors = getGraphQLErrorsFromResult(result);
   return isNonEmptyArray(errors);
-}
-
-export function getGraphQLErrorsFromResult<T>(result: FetchResult<T>) {
-  const graphQLErrors =
-    isNonEmptyArray(result.errors) ? result.errors.slice(0) : [];
-
-  if (
-    isExecutionPatchIncrementalResult(result) &&
-    isNonEmptyArray(result.incremental)
-  ) {
-    result.incremental.forEach((incrementalResult) => {
-      if (incrementalResult.errors) {
-        graphQLErrors.push(...incrementalResult.errors);
-      }
-    });
-  }
-  return graphQLErrors;
 }
